scripts/pyproject.toml
[tool.poetry]
name = "clear-sans-fonts"
version = "0.1.0"
description = "Scripts for Clear Sans fonts"
authors = ["Sean Leather"]
license = "BlueOak-1.0.0"

[tool.poetry.dependencies]
defcon = "^0.7.2"
python = "^3.6"
ufo-extractor = "^0.3.0"
ufonormalizer = "^0.4.2"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
